fix: Correct platformio.ini to resolve the right platform
Fixes platformio.ini configuration to ensure the correct pioarduino
platform is resolved consistently across all development environments.

## Platform Configuration
- Explicitly declare pioarduino platform (was causing resolution issues)
- Use Git version of Arduino_GFX library for proper compatibility
- Remove explicit framework version pinning (let pioarduino control versions)
- Remove C++ standard override (handled by pioarduino platform)

## Documentation & Troubleshooting
- Add comprehensive troubleshooting guide for Windows development setup
- Document pioarduino platform usage in build instructions

## Bug Fixes
- Fix format specifier warning for int32_t in debug output

This ensures PlatformIO resolves the correct platform dependencies
instead of relying on potentially inconsistent local cache.

+++ src/ui/ui_manager.cpp
@@ -743,7 +743,7 @@ void UIManager::update_settings_state() {
     char leight_sensor_info[256];
     snprintf(leight_sensor_info, sizeof(leight_sensor_info), 
         "Instant: %.2fg\nSmoothed: %.2fg\nSamples: %d\nRaw: %ld", 
-        weight_instant, weight_smoothed, sample_count, raw_reading);
+        weight_instant, weight_smoothed, sample_count, (long)raw_reading);
     
     unsigned long uptime_ms = millis();
     size_t free_heap = ESP.getFreeHeap();
